[#3428] [#3392]
- A8 format in Cairo was a "level of gray" type of image. Changed to ARGB32
The choice of ARGB32 instead of RGB24 was based on the "stride" requirement
in cairo. Reformatting the RGB24 buffer from gstreamer to respect this
would have been probably more expensive than that.

- Fixed a segmentation fault that originating from a bad shm_open in the
client. This was a synchronization kind of issue. The server would truncate
after the client has attached, resulting in segmentation fault. The client
now checks if the shm has size 0 and wait until some specified amount of
time.