[#3712] Generating a=fmtp line from encoder parameters
A videotestsrc element is connected to an input-selector element that
receives input from the appsrc on its second pad. To obtain the parameters
of interest, the videotestsrc element is selected and the resulting buffer
at the sink gets inspected on its GstCaps.

If some video codecs requires such behavior, then it must override
GstEncoder::setVideoFormat() in order to call generateSdp() in it.